Ezxtracted command line in separate class
