+++ pom.xml
@@ -16,6 +16,7 @@
 		<commons.cli.version>1.4</commons.cli.version>
 		<commons.io.version>2.6</commons.io.version>
 		<jgit.version>5.1.3.201810200350-r</jgit.version>
+		<guava.version>23.6-jre</guava.version>
 	</properties>
 
 	<licenses>
@@ -64,7 +65,7 @@
 								<transformer
 									implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
 									<manifestEntries>
-										<Main-Class>com.kgal.packagebuilder.PackageBuilder</Main-Class>
+										<Main-Class>com.kgal.packagebuilder.PackageBuilderCommandLine</Main-Class>
 									</manifestEntries>
 								</transformer>
 							</transformers>
@@ -123,7 +124,7 @@
 		<dependency>
 			<groupId>com.google.guava</groupId>
 			<artifactId>guava</artifactId>
-			<version>23.6-jre</version>
+			<version>${guava.version}</version>
 		</dependency>
 
 	</dependencies>
